added useNoteResizePanel hook
